Split test_ntheory.py
Split the number theory tests into files which more closely match the project's structure
